+++ mdps_ds_lib/stage_in_out/download_granules_factory.py
@@ -2,6 +2,7 @@ class DownloadGranulesFactory:
     S3 = 'S3'
     HTTP = 'HTTP'
     DAAC = 'DAAC'
+    AMALGAMATION = 'AMALGAMATION'
 
     def get_class(self, search_type):
         if search_type == DownloadGranulesFactory.S3:
@@ -13,4 +14,7 @@ class DownloadGranulesFactory:
         elif search_type == DownloadGranulesFactory.HTTP:
             from mdps_ds_lib.stage_in_out.download_granules_http import DownloadGranulesHttp
             return DownloadGranulesHttp()
+        elif search_type == DownloadGranulesFactory.AMALGAMATION:
+            from mdps_ds_lib.stage_in_out.download_granules_amalgamation import DownloadGranulesAmalgamation
+            return DownloadGranulesAmalgamation()
         raise ValueError(f'unknown search_type: {search_type}')